@@ -1,6 +1,6 @@
 {
   "name": "api-partners",
-  "version": "2.1.2",
+  "version": "2.1.8",
   "description": "CHUMS Partners Integration API",
   "main": "index.ts",
   "scripts": {
@@ -14,36 +14,38 @@
     "body-parser": "^1.19.0",
     "camelcase": "^6.3.0",
     "chums-local-modules": "github:UtahGooner/chums-local-modules",
+    "chums-types": "github:UtahGooner/chums-types",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.5",
     "csv-parse": "^5.0.3",
     "csvtojson": "^2.0.10",
     "date-fns": "^2.16.1",
+    "dayjs": "^1.11.7",
     "debug": "^4.3.1",
     "decimal.js": "^10.3.1",
     "dotenv": "^16.0.0",
     "express": "^4.17.1",
     "formidable": "^2.0.1",
-    "helmet": "^5.0.1",
-    "jsonwebtoken": "^8.5.1",
-    "mysql2": "^2.2.5",
+    "helmet": "^6.1.5",
+    "jsonwebtoken": "^9.0.0",
+    "mysql2": "^3.2.4",
     "node-fetch": "^2.6.6",
     "nodemailer": "^6.4.17",
     "pug": "^3.0.0",
-    "xml2js": "^0.4.23"
+    "xml2js": "^0.5.0"
   },
   "devDependencies": {
+    "@tsconfig/node18": "^2.0.0",
     "@types/body-parser": "^1.19.2",
     "@types/compression": "^1.7.2",
     "@types/cookie-parser": "^1.4.3",
-    "@types/csvtojson": "^2.0.0",
     "@types/debug": "^4.1.7",
     "@types/express": "^4.17.13",
     "@types/formidable": "^2.0.3",
     "@types/node-fetch": "^2.5.10",
     "@types/xml2js": "^0.4.8",
     "auto-changelog": "^2.4.0",
-    "typescript": "^4.6.3"
+    "typescript": "^5.0.4"
   },
   "auto-changelog": {
     "template": "keepachangelog",
