@@ -1,14 +1,11 @@
 "use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
 Object.defineProperty(exports, "__esModule", { value: true });
 const express_1 = require("express");
-const router = (0, express_1.Router)();
 const orders_1 = require("./orders");
-const products_1 = __importDefault(require("./products"));
-const product_feed_1 = __importDefault(require("./product-feed"));
+const products_1 = require("./products");
+const product_feed_1 = require("./product-feed");
 const feed_1 = require("./feed");
+const router = (0, express_1.Router)();
 router.get('/order/db/:ID(\\d+)/:format', orders_1.doLoadOrderFromDB);
 router.get('/orders', orders_1.doListOrders);
 router.get('/orders/since/:CreatedAfter(\\d+)', orders_1.doListOrders);
@@ -25,11 +22,11 @@ router.get('/order/ack/:AmazonOrderId', orders_1.doSubmitFeed_OrderAcknowledgeme
 router.get('/order/fulfill/:AmazonOrderId', orders_1.doSubmitFeed_OrderFulfillment);
 router.get('/order/onestep/:AmazonOrderId', orders_1.getOneStepOrder);
 router.post('/order/onestep/:AmazonOrderId', orders_1.getOneStepOrder);
-router.get('/products', products_1.default.getAvailable);
-router.post('/product', products_1.default.postProduct);
-router.get('/product/pricing/:SKU', products_1.default.getProductCompetitivePricing);
-router.get('/product/:ASIN', products_1.default.getProduct);
-router.get('/product-feed', product_feed_1.default.postFeed);
-router.post('/product-feed', product_feed_1.default.postFeed);
+router.get('/products', products_1.getAvailable);
+router.post('/product', products_1.postProduct);
+router.get('/product/pricing/:SKU', products_1.getProductCompetitivePricing);
+router.get('/product/:ASIN', products_1.getProduct);
+router.get('/product-feed', product_feed_1.postFeed);
+router.post('/product-feed', product_feed_1.postFeed);
 router.get('/feed/result/:FeedSubmissionId', feed_1.doGetFeedSubmissionResult);
 exports.default = router;
