@@ -173,8 +173,8 @@ async function loadFBMOrders(poList) {
                               oh.TaxableAmt + oh.NonTaxableAmt - oh.DiscountAmt`;
         const [rows] = await chums_local_modules_1.mysql2Pool.query(sql, { poList });
         rows.forEach(row => {
-            row.OrderTotal = new decimal_js_1.default(row.OrderTotal);
-            row.settlementTotal = new decimal_js_1.default(row.settlementTotal);
+            row.OrderTotal = new decimal_js_1.default(row.OrderTotal).toString();
+            row.settlementTotal = new decimal_js_1.default(row.settlementTotal).toString();
         });
         return rows;
     }
