@@ -129,8 +129,7 @@ async function convertToOrder(lines) {
         if (!header) {
             return Promise.reject('Unable to find Header Line');
         }
-        const customers = await (0, mapping_js_1.loadMatchingCustomers)(header);
-        const customer = customers.length === 1 ? customers[0] : null;
+        const customer = await (0, mapping_js_1.loadCustomer)(header);
         let mapping = [];
         const so = { ...defaultSalesOrder };
         if (customer && customer.ARDivisionNo) {
@@ -264,7 +263,6 @@ async function convertToOrder(lines) {
             SalesOrder,
             mapping,
             customer,
-            customers,
             unitsOfMeasure,
             ItemCodes
         };
