@@ -3,39 +3,33 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.postCustomer = exports.getCustomers = exports.deleteMapping = exports.postMapping = exports.getMapping = exports.loadShipToAddress = exports.loadBillToAddress = exports.loadItemUnits = exports.addCustomerMapping = exports.loadCustomerMapping = exports.loadMatchingCustomers = void 0;
+exports.postCustomer = exports.getCustomers = exports.deleteMapping = exports.postMapping = exports.getMapping = exports.loadShipToAddress = exports.loadBillToAddress = exports.loadItemUnits = exports.addCustomerMapping = exports.loadCustomerMapping = exports.loadCustomer = void 0;
 const debug_1 = __importDefault(require("debug"));
 const chums_local_modules_1 = require("chums-local-modules");
 const debug = (0, debug_1.default)('chums:lib:sps:mapping');
-async function loadMatchingCustomers(header) {
+async function loadCustomer(header) {
     try {
         const customers = await loadCustomers();
-        return customers.filter(customer => {
-            return customer.LookupFields.reduce((pv, cv) => {
-                if (cv.value.startsWith('^') || cv.value.endsWith('$')) {
-                    let re = /^INVALID_VALUE$/;
-                    try {
-                        re = new RegExp(cv.value);
-                    }
-                    catch (err) {
-                        re = /^INVALID_VALUE$/;
-                    }
-                    return pv && re.test(header[cv.field].toString());
-                }
-                return pv && header[cv.field] === cv.value;
-            }, true);
+        const [customer] = customers.filter(customer => {
+            let match = true;
+            customer.LookupFields
+                .forEach(({ field, value }) => {
+                match = match && header[field] === value;
+            });
+            return match;
         });
+        return customer ?? null;
     }
     catch (err) {
         if (err instanceof Error) {
-            console.debug("loadMatchingCustomers()", err.message);
+            console.debug("loadCustomer()", err.message);
             return Promise.reject(err);
         }
-        console.debug("loadMatchingCustomers()", err);
-        return Promise.reject(new Error('Error in loadMatchingCustomers()'));
+        console.debug("loadCustomer()", err);
+        return Promise.reject(new Error('Error in loadCustomer()'));
     }
 }
-exports.loadMatchingCustomers = loadMatchingCustomers;
+exports.loadCustomer = loadCustomer;
 async function loadCustomers() {
     try {
         const query = `SELECT c.id,
@@ -48,8 +42,7 @@ async function loadCustomers() {
                                       c.options
                                FROM sps_edi.customers c
                                         INNER JOIN c2.ar_customer ar
-                                                   USING (Company, ARDivisionNo, CustomerNo)
-                               ORDER BY JSON_LENGTH(c.LookupFields) desc`;
+                                                   USING (Company, ARDivisionNo, CustomerNo)`;
         const [rows] = await chums_local_modules_1.mysql2Pool.query(query);
         return rows.map((row) => {
             return {
